Cast size down.

+++ Pulp/3DEngine/Anim/Anim.cpp
@@ -21,7 +21,7 @@ void Anim::update(core::TimeVal delta, AnimState& state, Mat44Arr& bonesOut) con
 	// so we want generate data and shove it in the bones out.
 	// we have state to know last position.
 	// and delta since we last updated.
-	X_ASSERT(numBones() == bonesOut.size(), "Size mismatch")(numBones(), bonesOut.size());
+	X_ASSERT(numBones() == static_cast<int32_t>(bonesOut.size()), "Size mismatch")(numBones(), bonesOut.size());
 	
 	X_UNUSED(delta, state);
